migrations: if exists table

+++ database/migrations/2019_11_27_145947_create_devices_table.php
@@ -13,16 +13,18 @@ class CreateDevicesTable extends Migration
      */
     public function up()
     {
-        Schema::create('devices', function (Blueprint $table) {
-            $table->increments('id');
-            $table->string('ip_address', 15);
-            $table->string('description');
-            $table->unsignedInteger('type');
-            $table->boolean('active');
+        if (!Schema::hasTable('devices')) {
+            Schema::create('devices', function (Blueprint $table) {
+                $table->increments('id');
+                $table->string('ip_address', 15);
+                $table->string('description');
+                $table->unsignedInteger('type');
+                $table->boolean('active');
 
-            $table->foreign('type')->references('id')->on('devtypes')
-                ->onUpdate('cascade')->onDelete('cascade');
-        });
+                $table->foreign('type')->references('id')->on('devtypes')
+                    ->onUpdate('cascade')->onDelete('cascade');
+            });
+        }
     }
 
     /**
