+++ database/migrations/2019_11_27_151227_create_methods_table.php
@@ -13,19 +13,21 @@ class CreateMethodsTable extends Migration
      */
     public function up()
     {
-        Schema::create('methods', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('id_object');
-            $table->string('name', 100)->comment('Название метода объекта');
-            $table->string('easy', 20)->nullable()->comment('выполнение простого действия (например переключение порта). В значениях указываем номер порта устройства');
-            $table->unsignedInteger('script')->nullable()->comment('id скрипта из таблицы скриптов');
-            $table->string('comment');
+        if (!Schema::hasTable('methods')) {
+            Schema::create('methods', function (Blueprint $table) {
+                $table->increments('id');
+                $table->unsignedInteger('id_object');
+                $table->string('name', 100)->comment('Название метода объекта');
+                $table->string('easy', 20)->nullable()->comment('выполнение простого действия (например переключение порта). В значениях указываем номер порта устройства');
+                $table->unsignedInteger('script')->nullable()->comment('id скрипта из таблицы скриптов');
+                $table->string('comment');
 
-            $table->foreign('id_object')->references('id')->on('objects')
-                ->onUpdate('cascade')->onDelete('cascade');
-            $table->foreign('script')->references('id')->on('scripts')
-                ->onUpdate('cascade')->onDelete('set null');
-        });
+                $table->foreign('id_object')->references('id')->on('objects')
+                    ->onUpdate('cascade')->onDelete('cascade');
+                $table->foreign('script')->references('id')->on('scripts')
+                    ->onUpdate('cascade')->onDelete('set null');
+            });
+        }
     }
 
     /**
