+++ database/migrations/2019_11_27_151447_create_graph_termostats_table.php
@@ -13,15 +13,17 @@ class CreateGraphTermostatsTable extends Migration
      */
     public function up()
     {
-        Schema::create('graph_termostats', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('id_termostat')->comment('id термостата из таблицы термостатов');
-            $table->dateTime('datetime')->comment('дата и время значения');
-            $table->double('value')->comment('значение параметра');
+        if (!Schema::hasTable('graph_termostats')) {
+            Schema::create('graph_termostats', function (Blueprint $table) {
+                $table->increments('id');
+                $table->unsignedInteger('id_termostat')->comment('id термостата из таблицы термостатов');
+                $table->dateTime('datetime')->comment('дата и время значения');
+                $table->double('value')->comment('значение параметра');
 
-            $table->foreign('id_termostat')->references('id')->on('termostats')
-                ->onUpdate('cascade')->onDelete('cascade');
-        });
+                $table->foreign('id_termostat')->references('id')->on('termostats')
+                    ->onUpdate('cascade')->onDelete('cascade');
+            });
+        }
     }
 
     /**
